feat: VSDSPUB-59 Extracting Client logic to separate module
