+++ ldes-client/src/main/java/be/vlaanderen/informatievlaanderen/ldes/client/models/TreeDirection.java
@@ -1,0 +1,5 @@
+package be.vlaanderen.informatievlaanderen.ldes.client.models;
+
+public enum TreeDirection {
+    GreaterThanRelation, LessThanRelation
+}
