+++ ldes-client/src/main/java/be/vlaanderen/informatievlaanderen/ldes/client/services/LdesService.java
@@ -1,0 +1,10 @@
+package be.vlaanderen.informatievlaanderen.ldes.client.services;
+
+import java.util.List;
+
+public interface LdesService {
+
+    boolean hasPagesToProcess();
+    String getNextPageUrl();
+    List<String[]> processNextPage();
+}
