+++ ldes-processors/src/main/java/be/vlaanderen/informatievlaanderen/ldes/processors/config/LdesProcessorRelationships.java
@@ -1,0 +1,10 @@
+package be.vlaanderen.informatievlaanderen.ldes.client.config;
+
+import org.apache.nifi.processor.Relationship;
+
+public class LdesProcessorRelationships {
+    public static final Relationship DATA_RELATIONSHIP = new Relationship.Builder()
+            .name("data")
+            .description("LDES item updates")
+            .build();
+}
