--- nifi-processors/src/main/java/be/vlaanderen/vsds/ldesclient/nifi/processors/config/LdesProcessorRelationships.java
@@ -1,10 +1,0 @@
-package be.vlaanderen.vsds.ldesclient.nifi.processors.config;
-
-import org.apache.nifi.processor.Relationship;
-
-public class LdesProcessorRelationships {
-    public static final Relationship DATA_RELATIONSHIP = new Relationship.Builder()
-            .name("data")
-            .description("LDES item updates")
-            .build();
-}
