--- nifi-processors/src/main/java/be/vlaanderen/vsds/ldesclient/nifi/processors/models/TreeDirection.java
@@ -1,5 +1,0 @@
-package be.vlaanderen.vsds.ldesclient.nifi.processors.models;
-
-public enum TreeDirection {
-    GreaterThanRelation, LessThanRelation
-}
