+++ pom.xml
@@ -4,14 +4,15 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>be.vlaanderen.informatievlaanderen.ldes-client</groupId>
-    <artifactId>nifi-processors-bundle</artifactId>
+    <groupId>be.vlaanderen.informatievlaanderen</groupId>
+    <artifactId>ldes-processors-bundle</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>nifi-processors</module>
-        <module>nifi-processors-nar</module>
+        <module>ldes-client</module>
+        <module>ldes-processors</module>
     </modules>
+
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
